SPI and MAX6675 functional Library

+++ Firmware/A10102_V1.0/A10102_V1.0/src/MAX6675.h
@@ -14,6 +14,6 @@
 
 #include "SPI.h"
 
-uint16_t Temperature_Read(uint8_t spi_read, uint8_t spi_slave);
+float Temperature_Read(uint8_t spi_read, uint8_t spi_slave);
 
 #endif /* MAX6675_H_ */
